+++ api/src/test/java/org/grnet/cat/api/AssessmentsEndpointTest.java
@@ -496,11 +496,13 @@ public class AssessmentsEndpointTest extends KeycloakTest {
         register("validated");
         register("admin");
 
-        makeValidation("validated", 6L);
-        fetchTemplateByActorAndType();
+        var validation = makeValidation("validated", 6L);
+        var templateDto = fetchTemplateByActorAndType();
 
         var request = new JsonAssessmentRequest();
-        request.assessmentDoc = makeJsonDoc(false, 6L);
+        request.validationId = validation.id;
+        request.templateId = templateDto.id;
+        request.assessmentDoc = makeJsonDoc(false);
 
         var assessment = given()
                 .auth()
@@ -534,11 +536,13 @@ public class AssessmentsEndpointTest extends KeycloakTest {
         register("validated");
         register("admin");
 
-        makeValidation("validated", 6L);
-        fetchTemplateByActorAndType();
+        var validation = makeValidation("validated", 6L);
+        var templateDto = fetchTemplateByActorAndType();
 
         var request = new JsonAssessmentRequest();
-        request.assessmentDoc = makeJsonDoc(true, 6L);
+        request.validationId = validation.id;
+        request.templateId = templateDto.id;
+        request.assessmentDoc = makeJsonDoc(true);
 
         var assessment = given()
                 .auth()
@@ -563,7 +567,7 @@ public class AssessmentsEndpointTest extends KeycloakTest {
                 .extract()
                 .as(InformativeResponse.class);
 
-        assertEquals("It is not allowed to manage a published assessment.", informativeResponse.message);
+        assertEquals("It is not allowed to delete a published assessment.", informativeResponse.message);
     }
 
     @Test
@@ -571,13 +575,15 @@ public class AssessmentsEndpointTest extends KeycloakTest {
 
         register("validated");
         register("admin");
-        register("bob");
+        register("evald");
 
-        makeValidation("validated", 6L);
-        fetchTemplateByActorAndType();
+        var validation = makeValidation("validated", 6L);
+        var templateDto = fetchTemplateByActorAndType();
 
         var request = new JsonAssessmentRequest();
-        request.assessmentDoc = makeJsonDoc(true, 6L);
+        request.validationId = validation.id;
+        request.templateId = templateDto.id;
+        request.assessmentDoc = makeJsonDoc(true);
 
         var assessment = given()
                 .auth()
@@ -593,7 +599,7 @@ public class AssessmentsEndpointTest extends KeycloakTest {
 
         var informativeResponse = given()
                 .auth()
-                .oauth2(getAccessToken("bob"))
+                .oauth2(getAccessToken("evald"))
                 .contentType(ContentType.JSON)
                 .delete("/{id}", assessment.id)
                 .then()
@@ -602,7 +608,7 @@ public class AssessmentsEndpointTest extends KeycloakTest {
                 .extract()
                 .as(InformativeResponse.class);
 
-        assertEquals("User not authorized to manage assessment with ID "+assessment.id, informativeResponse.message);
+        assertEquals("User not authorized to delete assessment with ID "+assessment.id, informativeResponse.message);
     }
 
     private ValidationResponse makeValidation(String username, Long actorId) {